refactor(projectSubmissions): reloacate and rename folder/files
