+++ src/frontend/src/components/ProjectSubmissions/Infographics/InfographicsCard.tsx
