+++ src/frontend/src/components/ProjectSubmissions/Infographics/TaskSubmissionsMapLegend.tsx
